Move GitHub tasks to Celery worker
