+++ openedx_webhooks/templates/github_rescan.html
@@ -24,9 +24,7 @@
     </p>
     <p>Special values this form accepts are:</p>
     <ul>
-      <li> "all": scans all repos except for edx-platform.</li>
-      <li>"all.1": scans the first half of repos, except edx-platform.</li>
-      <li>"all.2": scans the second half of repos, except edx-platform.</li>
+      <li>"all": scans all repos in the repos.yaml file</li>
     </ul>
     <label for="repo">Repo to scan</label>
     <input type="text" name="repo" id="repo" value="edx/edx-platform" />
